+++ src/algo/tiger.h
@@ -37,7 +37,7 @@ typedef struct {
     uint64_t w[8];
     uint8_t  t[8];
     uint8_t  block[2 * AKMOS_TIGER_BLKLEN];
-    uint32_t total;
+    uint64_t total;
     size_t   len;
 } akmos_tiger_t;
